+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 dev:
   - speed up sync by only updating changed validators
   - add open telemetry tracing
+  - avoid fetching duplicate proposer duties
 
 0.6.17:
   - expose read-only transaction functions